requirements.txt
absl-py==2.2.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
astunparse==1.6.3
attrs==25.3.0
backtrader==1.9.78.123
bcrypt==4.3.0
beautifulsoup4==4.13.3
bidict==0.23.1
blinker==1.9.0
blis==1.3.0
catalogue==2.0.10
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.21.0
cloudscraper==1.2.71
colorama==0.4.6
confection==0.1.5
contourpy==1.3.1
cycler==0.12.1
cymem==2.0.11
dateparser==1.2.1
DateTime==5.5
dnspython==1.16.0
fastapi==0.115.12
feedparser==6.0.11
filelock==3.18.0
Flask==3.1.0
flask-cors==5.0.1
Flask-SocketIO==5.5.1
flatbuffers==25.2.10
fonttools==4.56.0
frozendict==2.4.6
frozenlist==1.6.0
fsspec==2025.3.2
fuzzywuzzy==0.18.0
gast==0.6.0
google-pasta==0.2.0
greenlet==3.2.1
grpcio==1.71.0
h11==0.14.0
h5py==3.13.0
huggingface-hub==0.30.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
keras==3.9.0
kiwisolver==1.4.8
langcodes==3.5.0
langdetect==1.0.9
language_data==1.3.0
Levenshtein==0.27.1
libclang==18.1.1
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
MetaTrader5==5.0.4874
ml_dtypes==0.5.1
mplfinance==0.12.10b0
mpmath==1.3.0
multidict==6.4.3
multitasking==0.0.11
murmurhash==1.0.12
namex==0.0.8
narwhals==1.29.1
networkx==3.4.2
nltk==3.9.1
numpy==2.1.3
opt_einsum==3.4.0
optree==0.14.1
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pandas_ta==0.3.14b0
patsy==1.0.1
peewee==3.17.9
pillow==11.1.0
platformdirs==4.3.7
playwright==1.51.0
plotly==6.0.0
polygon-api-client==1.14.4
preshed==3.0.9
propcache==0.3.1
protobuf==5.29.4
pycountry==24.6.1
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.10.6
pydantic_core==2.27.2
pyee==12.1.1
Pygments==2.19.1
PyJWT==2.10.1
pymongo==3.11.0
pyparsing==3.2.1
PySocks==1.7.1
python-binance==1.0.28
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-engineio==4.12.0
python-http-client==3.3.7
python-Levenshtein==0.27.1
python-socketio==5.13.0
pytz==2025.1
PyYAML==6.0.2
RapidFuzz==3.13.0
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
safetensors==0.5.3
schedule==1.2.2
scikit-learn==1.6.1
scipy==1.15.2
selenium==4.31.0
sendgrid==6.11.0
sgmllib3k==1.0.0
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
spacy==3.8.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
starkbank-ecdsa==2.2.0
starlette==0.46.1
statsmodels==0.14.4
sympy==1.13.1
ta==0.11.0
ta-lib @ file:///C:/Users/lenovo/Downloads/ta_lib-0.6.3-cp311-cp311-win_amd64.whl#sha256=c3a4e29e7e97dbd03c8d8bef0d26da3ced6cc34cb6af56134d273a0284c748d8
tabulate==0.9.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_intel==2.18.0
termcolor==2.5.0
tf_keras==2.19.0
thinc==8.3.6
threadpoolctl==3.5.0
tokenizers==0.21.1
torch==2.6.0
tqdm==4.67.1
transformers==4.51.0
trio==0.29.0
trio-websocket==0.12.2
typer==0.15.3
typing_extensions==4.13.0
tzdata==2025.1
tzlocal==5.3.1
undetected-chromedriver==3.5.5
urllib3==2.3.0
uvicorn==0.34.0
wasabi==1.1.3
waybackpy==3.0.6
weasel==0.4.1
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==14.2
Werkzeug==3.1.3
wrapt==1.17.2
wsproto==1.2.0
yarl==1.20.0
yfinance==0.2.55
zope.interface==7.2
